intcode day 7

+++ 2019/.skel/src/main.rs
@@ -5,10 +5,10 @@ fn main() {
     println!("Part 1: {}", p1(&inp));
     println!("Part 2: {}", p2(&inp));
 }
-fn p1(inp: &str) -> i32{
+fn p1(inp: &str) -> isize{
     return 0;
 }
-fn p2(inp: &str) -> i32{
+fn p2(inp: &str) -> isize{
     return 0;
 }
